Fixing reverted changes and moving auth to api folder
