--- src/db/database.ts
@@ -1,6 +1,0 @@
-import knex from "knex";
-import config from "./knex/knexfile";
-
-const db = knex(config);
-
-export default db;
