+++ src/db/knex.connector.ts
@@ -1,0 +1,6 @@
+import knex from "knex";
+import config from "./knex/knexfile";
+
+const knexConnector = knex(config);
+
+export default knexConnector;
