+++ src/repositories/impl/user.repository.impl.ts
@@ -1,11 +1,11 @@
-import knex_connector from "../../db/knex.connector";
+import knexConnector from "../../db/knex.connector";
 import { CreateUserDto } from "../../dtos/user/create-user.dto";
 import User from "../../model/user";
 import { Repository } from "../entity.repository";
 
 export class UserRepositoryImpl implements Repository<CreateUserDto, User> {
   constructor() {
-    User.knex(knex_connector);
+    User.knex(knexConnector);
   }
 
   public async insert(user: User): Promise<User> {
